refactor : 로또 구매 가격 상수 타입 변경

+++ src/main/java/lotto/domain/Money.java
@@ -2,8 +2,8 @@ package lotto.domain;
 
 public class Money {
     private long amount;
-    private static final long MINIMUM_LOTTO_PRICE = 1000;
-    private static final long LOTTO_PRICE_UNIT = 1000;
+    private static final int MINIMUM_LOTTO_PRICE = 1000;
+    private static final int LOTTO_PRICE_UNIT = 1000;
 
     public Money(long money) {
         validateMoney(money);
